clean-up host code
